add start logic and stop logic

+++ cmd/start.go
@@ -1,29 +1,15 @@
-// Copyright © 2018 NAME HERE <EMAIL ADDRESS>
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
 package cmd
 
 import (
-	"encoding/json"
 	"fmt"
+	"github.com/garyburd/redigo/redis"
 	"github.com/spf13/cobra"
 	"github.com/spf13/viper"
 	"io/ioutil"
-	"log"
 	"net/http"
 	"os"
 	"os/exec"
+	"socket/api"
 	"socket/database"
 	"socket/internal/im"
 )
@@ -34,21 +20,25 @@ var daemon bool
 var startCmd = &cobra.Command{
 	Use:   "start",
 	Short: "Start Web Socket service",
-	Long: "Start Web Socket service on this host",
+	Long:  "Start Web Socket service on this host",
 	Run: func(cmd *cobra.Command, args []string) {
 		if daemon {
+			exist := isExist("socket.lock")
+			if exist {
+				fmt.Println("Service is started.")
+				os.Exit(0)
+			}
 			command := exec.Command("socket", "start")
 			if err := command.Start(); err != nil {
-				panic(err)
+				fmt.Println("Service startup failed.")
+				os.Exit(0)
 			}
-			fmt.Printf("socket start, [PID] %d running...\n", command.Process.Pid)
+			fmt.Println("Service started successfully.")
 			if err := ioutil.WriteFile("socket.lock", []byte(fmt.Sprintf("%d", command.Process.Pid)), 0666); err != nil {
 				panic(err)
 			}
 			daemon = false
 			os.Exit(0)
-		} else {
-
 		}
 		startService()
 	},
@@ -69,47 +59,39 @@ func init() {
 	startCmd.Flags().BoolVarP(&daemon, "daemon", "d", false, "Start the service as a daemon")
 }
 
-func startService()  {
+func startService() {
+	bootstrap()
+	api.Register()
 	web := viper.GetStringMapString("web")
-	api := viper.GetStringMapString("api")
-	redis := viper.GetStringMapString("redis")
 	socket := viper.GetStringMapString("socket")
-	prefix := api["prefix"]
-
-	if prefix == "" {
-		prefix = "/api"
-	}
-
-	type Response struct {
-		Status int `json:"status"`
-		Message string `json:"message"`
-		Data interface{}
-	}
-
-	http.HandleFunc(prefix, func(writer http.ResponseWriter, request *http.Request) {
-		data := make([]string, 0)
-		response, err := json.Marshal(&Response{
-			Status: 200,
-			Message: "Success",
-			Data: data,
-		})
-		if err != nil {
-			log.Fatal(err)
-		}
-		if _, err := writer.Write(response); err != nil {
-			log.Fatal(err)
-		}
-	})
-
-	redisAddr := fmt.Sprintf("%v:%v", redis["host"], redis["port"])
-	database.CreateRedisPool(redisAddr, redis["pass"])
 	http.HandleFunc(socket["prefix"], im.Handle)
 	http.Handle("/", http.FileServer(http.Dir("web")))
 	socketAddr := fmt.Sprintf("%v:%v", web["host"], web["port"])
 
-	fmt.Printf("Web service listen on %v", socketAddr)
+	fmt.Printf("Web service listen on %v \n", socketAddr)
 	if err := http.ListenAndServe(socketAddr, nil); err != nil {
-		log.Fatal("ListenAndServe: ", err)
+		fmt.Println(err)
 	}
 }
 
+func bootstrap() {
+	redisConf := viper.GetStringMapString("redis")
+	redisAddr := fmt.Sprintf("%v:%v", redisConf["host"], redisConf["port"])
+	database.CreateRedisPool(redisAddr, redisConf["pass"])
+	clear()
+}
+
+func clear() {
+	connection := database.Pool.Get()
+	if _, err := redis.String(connection.Do("FLUSHDB")); err != nil {
+		panic(err)
+	}
+	if err := connection.Close(); err != nil {
+		fmt.Println(err)
+	}
+}
+
+func isExist(f string) bool {
+	_, err := os.Stat(f)
+	return err == nil || os.IsExist(err)
+}
